RUB-293: split safe Go consensus shape rows
